CORDA-499: Rename CurrencyUtils to Currencies (#1476)
* Rename CurrencyUtils to Currencies

Rename CurrencyUtils to Currencies to explain why it's not in the finance utils package.

* Correct reference in CordaRPCJavaClientTest

* Remove spurious newline
