+++ finance/src/main/kotlin/net/corda/finance/Currencies.kt
@@ -1,4 +1,4 @@
-@file:JvmName("CurrencyUtils")
+@file:JvmName("Currencies")
 
 package net.corda.finance
